migrate streaming logic to `BaseChatHandler`
